app-support: break out the thread backed `Backoff` impl
